requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cymem==2.0.2
cytoolz==0.9.0.1
dash==0.37.0
dash-core-components==0.43.1
dash-html-components==0.13.5
dash-renderer==0.18.0
dash-table==3.4.0
decorator==4.3.2
dill==0.2.9
Flask==1.0.2
Flask-Compress==1.4.0
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbformat==4.4.0
numpy==1.16.1
pandas==0.24.1
Pillow==5.4.1
plac==0.9.6
plotly==3.6.1
preshed==2.0.1
python-dateutil==2.8.0
pytz==2018.9
regex==2018.1.10
requests==2.21.0
retrying==1.3.3
six==1.12.0
spacy==2.0.18
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
traitlets==4.3.2
ujson==1.35
urllib3==1.24.1
Werkzeug==0.14.1
wordcloud==1.5.0
wrapt==1.10.11
